Refactor package location of actions and criteria
