+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/CVImageCriteriaEvaluator.cs
@@ -3,8 +3,8 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using RegressionGames.StateRecorder.BotSegments.Models;
 using RegressionGames.StateRecorder.BotSegments.Models.AIService;
+using RegressionGames.StateRecorder.BotSegments.Models.BotCriteria;
 using RegressionGames.StateRecorder.JsonConverters;
 using StateRecorder.BotSegments;
 using UnityEngine;
