+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/CVTextCriteriaEvaluator.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using RegressionGames.StateRecorder.BotSegments.Models;
 using RegressionGames.StateRecorder.BotSegments.Models.AIService;
+using RegressionGames.StateRecorder.BotSegments.Models.BotActions;
+using RegressionGames.StateRecorder.BotSegments.Models.BotCriteria;
 using UnityEngine;
 
 namespace RegressionGames.StateRecorder.BotSegments
