+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/JsonConverters/BehaviourActionDataJsonConverter.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using RegressionGames.StateRecorder.BotSegments.Models;
-using UnityEngine;
+using RegressionGames.StateRecorder.BotSegments.Models.BotActions;
 
 namespace RegressionGames.StateRecorder.BotSegments.JsonConverters
 {
