+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/JsonConverters/CVImageMouseActionDataJsonConverter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using RegressionGames.StateRecorder.BotSegments.Models;
+using RegressionGames.StateRecorder.BotSegments.Models.BotActions;
 
 namespace RegressionGames.StateRecorder.BotSegments.JsonConverters
 {
