+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/JsonConverters/RandomMouseObjectActionDataJsonConverter.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using RegressionGames.StateRecorder.BotSegments.Models;
+using RegressionGames.StateRecorder.BotSegments.Models.BotActions;
 using UnityEngine;
 
 namespace RegressionGames.StateRecorder.BotSegments.JsonConverters
