+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotActions/RandomMouseObjectActionData.cs
@@ -7,7 +7,7 @@ using RegressionGames.StateRecorder.Models;
 using UnityEngine;
 using Random = UnityEngine.Random;
 
-namespace RegressionGames.StateRecorder.BotSegments.Models
+namespace RegressionGames.StateRecorder.BotSegments.Models.BotActions
 {
     /**
      * <summary>Data for clicking on a random renderable object in the frame</summary>
