+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotCriteria.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c930171fe9794c4db241172da63bfefc
+timeCreated: 1730903427
