+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotCriteria/ActionCompleteKeyFrameCriteriaData.cs
@@ -2,28 +2,28 @@ using System;
 using System.Text;
 using RegressionGames.StateRecorder.JsonConverters;
 
-namespace RegressionGames.StateRecorder.BotSegments.Models
+namespace RegressionGames.StateRecorder.BotSegments.Models.BotCriteria
 {
     [Serializable]
     public class ActionCompleteKeyFrameCriteriaData : IKeyFrameCriteriaData
     {
         // update this if this schema changes
         public int apiVersion = SdkApiVersion.VERSION_11;
-        
+
         public void WriteToStringBuilder(StringBuilder stringBuilder)
         {
             stringBuilder.Append("{\"apiVersion\":");
             IntJsonConverter.WriteToStringBuilder(stringBuilder, apiVersion);
             stringBuilder.Append("}");
         }
-        
+
         public override string ToString()
         {
             StringBuilder sb = new StringBuilder(100);
             WriteToStringBuilder(sb);
             return sb.ToString();
         }
-        
+
         public int EffectiveApiVersion()
         {
             return apiVersion;
