+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotCriteria/AndKeyFrameCriteriaData.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using RegressionGames.StateRecorder.JsonConverters;
 
-namespace RegressionGames.StateRecorder.BotSegments.Models
+namespace RegressionGames.StateRecorder.BotSegments.Models.BotCriteria
 {
     [Serializable]
     public class AndKeyFrameCriteriaData : IKeyFrameCriteriaData
