+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotCriteria/AndKeyFrameCriteriaData.cs.meta
