+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotCriteria/CVObjectDetectionKeyFrameCriteriaData.cs.meta
