+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotCriteria/OrKeyFrameCriteriaData.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using RegressionGames.StateRecorder.JsonConverters;
 
-namespace RegressionGames.StateRecorder.BotSegments.Models
+namespace RegressionGames.StateRecorder.BotSegments.Models.BotCriteria
 {
     [Serializable]
     public class OrKeyFrameCriteriaData : IKeyFrameCriteriaData
