+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotCriteria/PathKeyFrameCriteriaData.cs
@@ -2,7 +2,7 @@
 using System.Text;
 using RegressionGames.StateRecorder.JsonConverters;
 
-namespace RegressionGames.StateRecorder.BotSegments.Models
+namespace RegressionGames.StateRecorder.BotSegments.Models.BotCriteria
 {
     [Serializable]
     public class PathKeyFrameCriteriaData : IKeyFrameCriteriaData
