+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/ScreenRecorder.cs
@@ -13,6 +13,8 @@ using Newtonsoft.Json;
 using RegressionGames.CodeCoverage;
 using RegressionGames.RemoteOrchestration;
 using RegressionGames.StateRecorder.BotSegments.Models;
+using RegressionGames.StateRecorder.BotSegments.Models.BotActions;
+using RegressionGames.StateRecorder.BotSegments.Models.BotCriteria;
 using RegressionGames.StateRecorder.Models;
 using StateRecorder.BotSegments;
 #if UNITY_EDITOR
